/getMyStarredContent endpoint
